refactor: 색상 팔레트를 위한 styled component theme 적용
