+++ src/styles/theme.d.ts
@@ -1,0 +1,6 @@
+import 'styled-components';
+
+declare module 'styled-components' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  export interface DefaultTheme {}
+}
