+++ src/styles/theme.ts
@@ -1,0 +1,22 @@
+import { DefaultTheme } from 'styled-components';
+
+const color = {
+  white: '#fff',
+  gray100: '#f8f8f8',
+  gray200: '#e9e9e9',
+  gray300: '#d6d6d6',
+  gray400: '#727272',
+  gray500: '#525252',
+  gray600: '#363636',
+  black: '#333',
+
+  primary: '#1a7cff',
+
+  red: '#ce0000',
+};
+
+const theme: DefaultTheme = {
+  color,
+};
+
+export default theme;
